RA: o2cb: Optimize redistribution of cluster.cnf.
Instead of the leader pushing out changes, the slaves now compare the
md5sum from the leader (used as the magic lock value) to their own, and
only retrieve when the cluster.conf has actually changed.

This also allows for greater concurrency and might be an interesting
snippet of code for other RAs as well.