[LDIRECTORD] Use negotiatetimeout and connecttimeout as mutual defaults Use negotiatetimeout and connecttimeout as mutual defaults.
There is some confusion about when to use negotiatetimeout and when
to use connecttimeout. This should aleviate the problem a little
by using one as the default for the other. So if one is set, it will
become the value for both with the normal global/per-virtual scoping
rules honoured.

As suggested by Tuomo Soini

Another option is to combine negotiatetimeout and connecttimeout into
a single timeout, which I am happy to consider at a later date.

Signed-off-by: Simon Horman <horms@verge.net.au>