[ldirectord] fix sql select queries - at the expense of non-select unfortunately, the stuff to make the queries work with non-select doesn't actually work with anything doing a select out of the db, at least if you do something like select *
however, it is completely asinine to assume someone's going to regularly
insert data from the director into dbs he's loadbalancing across, so you
can rule that out pretty much.

I _CAN_ make this work if you want to do an insert / update in addition to
the select, but it doesn't necessarily make sense to me.

Cc: "mandrake (Geoff Harrison)" <mandrake@mandrake.net>
Signed-off-by: Simon Horman <horms@verge.net.au>