debian,ldirectord: add LSB formatted dependency info in init.d script To be able to check boot script order, and also to be able to start boot scripts in parallel, it is important to know the dependencies of the various boot scripts.  The Linux Software Base specifies a init.d header file format useful for this purpose, and adding such header to the ldirectord init.d script would make it possible for us to use this information to check the current sequence and to speed up the Debian boot.
I am helping out with a system to update the boot sequence based on
these dependencies, and would like see this as the default in Lenny.
Because of this, it is nice if the dependencies was updated quickly.

<URL:http://refspecs.freestandards.org/LSB_2.1.0/LSB-generic/LSB-generic/initscrcomconv.html>
documents the LSB header format.  Some Debian notes are available from
<URL:http://wiki.debian.org/LSBInitScripts>.


See Debian Bug: #469764
http://bugs.debian.org/469764

Signed-off-by: Simon Horman <horms@verge.net.au>