build: Don't distribute snmp_subagent As configure.in currently stands, as of the hg dev tree 54723736ab18, make dist fails because autogenerated files in snmp_subagent/, in parcicular snmp_subagent/Makefile.am, are not created.
Andrew Beekhof tells me that this is because snmp_subagent can't build
without the crm, which is now in the pacemaker tree along with
snmp_subagent.

This fairly minimal patch allows make dist to succeed by not
distributing snmp_subagent/. I have found that building the resulting
tarball also succeeds.

A fuller patch would either fix snmp_subagent so that it can compile,
or remove it completely from snmp_subagent/, configure.in and
heartbeat.rpm.spec. I don't feel that it is within my jurisdiction
to make either change at this time.

Signed-off-by: Simon Horman <horms@verge.net.au>