hb_report: PE inputs pattern not used anymore
In the previous changeset (4843283c073c) we dropped collecting PE
input files based on the log, because it that is also dependent
on the log format. They are collected based on their timestamps
now. Further advantage is that should we fail to find logs,
getting PE input files would still work.