medium: ldirectord: Ignore exit status of left-over external check processes Hi,
When using external check processes in ldirectord, we encountered some
cases of zombies being left behind when the external process ran into
a timeout. This happened because nobody wait()ed on their exit status.

The following patch fixes this issue for us by just ignoring
SIGCHLD signals in the code branch that executes the checks.

Julius

Signed-off-by: Simon Horman <horms@verge.net.au>