High: stonith bladehpi (bnc#389344): fix a mix of a threaded library and not threaded stonithd
The bladehpi stonith plugin is using openhpi which is using
glib resources with threads. Now, when stonithd wants to
manage glib objects, it ends up waiting forever for a mutex.
Resolved by moving all calls which are using threads in openhpi
to children of stonithd. That way we keep a good plugin and
stonithd sane.