ldirectord: Make quiescent=no the default
The main advantage of quiescent is that it allows the graceful
stop of services. However it seems likely that ldirectord is
predominantly used to handle the failure or real-servers, in
which case quiescent=no should actually work better as it will
drop connections, rather than continuing to forward them.

A further disadvantage of using quiescent=yes as the default
is that setting the weight of real servers to 0 and thus leaving
them in the LVS table seems likely to be more confusing to new
users than just removing them from the LVS table, which is what
quiescent=no.

The sample configuration has previously had quiescent=yes.
This change will not affect users that left quiescent=yes in
the configuration or set it to quiescent=no. It will
change the behaviour for users who commented out or removed
the quiescent directive from their ldirectord.cf. This seems
likely to be few people, if any.

Cc: Tuomo Soini <tis@foobar.fi>
Signed-off-by: Simon Horman <horms@verge.net.au>