Force resetting of unset configuration directives to defaults.
The default values for global configuration directives are set once
early on in the start up.  If a configuration file specifies a global
configuration option and then, at a later time, that option is removed
from the configuration file, the global variable will neither get set
nor reset during the configuration reload.  This results in the old setting
persisting, instead of the expect and documented "default" value in
the absence of the configuration directive.

This patch puts all the defaults into a function (set_defaults) which
is called once at startup and once at every configuration reload.  All
global configuration defaults should be set in this function when added
or changed.

Signed-Off-By: Sean E. Millichamp <sean@bruenor.org>
Signed-Off-By: Simon Horman <horms@verge.net.au>