Low: fix wrong configure check for clock_t
The test for CLOCK_T_IS_LONG_ENOUGH should have checked sizeof(clock_t),
not sizeof(long). This is relevant for clplumbing, which has been moved
to glue, and is being fixed now, tracked as
http://developerbugs.linux-foundation.org/show_bug.cgi?id=2596

As that define is used in <clplumbing/longclock.h>, we need to
adjust the configure test in heartbeat proper, too.