fix emergency shutdown due to broken update_ackseq
If we have an asymmetrical network problem (packets going through in only one
direction), we may think a node is alive (we see his packets), but he does not
send ACKs for our packets (he does not see them).

If the message queue history filled up before we saw the first ACK,
the "start" offset in update_ackseq was chosen wrong, an assert triggers,
and heartbeat goes into emergency shutdown.

Since the message queue history may have wrapped already,
we still need to start the cleanup with "lowseq", not a hardcoded 0.